[feat] add login inser User

+++ src/app/components/users-table/users-table.component.ts
@@ -3,8 +3,7 @@ import { MatSort } from '@angular/material/sort';
 import { MatDialog } from '@angular/material/dialog';
 import { UserService } from 'src/app/services/user.service';
 import { MatTableDataSource } from '@angular/material/table';
-import { Component, OnInit, ViewChild, ElementRef } from '@angular/core';
-import { debounceTime, distinctUntilChanged, pluck } from 'rxjs/operators';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { PaginatorLabels } from 'src/app/services/PaginatorLabels.service';
 import { MatPaginator, MatPaginatorIntl } from '@angular/material/paginator';
 import { alertService } from './../../shared/modules/alert/services/alert.service';
@@ -20,7 +19,7 @@ export class UsersTableComponent implements OnInit {
   @ViewChild(MatPaginator) paginator!: MatPaginator;
   @ViewChild(MatSort) sort!: MatSort;
   dataSource!: MatTableDataSource<User[]>;
-
+  private pending: boolean = false;
   constructor(
     public dialog: MatDialog,
     private readonly _userService: UserService,
@@ -36,14 +35,19 @@ export class UsersTableComponent implements OnInit {
   }
 
   private getUsersList(search: string = '') {
-    this._userService
-      .getUsers(search)
-      .pipe(debounceTime(1000), distinctUntilChanged())
-      .subscribe((res) => {
+    if (this.pending) return;
+    this.pending = true;
+    this._userService.getUsers(search).subscribe(
+      (res) => {
+        this.pending = false;
         this.dataSource = new MatTableDataSource(res);
         this.dataSource.paginator = this.paginator;
         this.dataSource.sort = this.sort;
-      });
+      },
+      (e) => {
+        this.pending = false;
+      }
+    );
   }
 
   onAddUser() {
@@ -54,7 +58,45 @@ export class UsersTableComponent implements OnInit {
         message: 'افزودن کاربر ',
         cancel: 'بستن',
         submit: 'ذخیره اطلاعات',
-        submitFn: () => {},
+        submitFn: (data: User) => {
+          this.createUser(data);
+          dialog.close();
+        },
+        cancelFn: () => {
+          dialog.close();
+        },
+      },
+    });
+  }
+
+  private createUser(data: User) {
+    if (this.pending) this.pending = true;
+    this._userService.createUser(data).subscribe(
+      (res) => {
+        this.pending = false;
+        this.getUsersList();
+      },
+      (e) => {
+        this.pending = false;
+        this.alertService.showSuccess(
+          ' خطایی هنگام افزودن کاربر جدی روی داده است '
+        );
+      }
+    );
+  }
+
+  editUserData(data: User) {
+    const dialog = this.dialog.open(InsertEditDialogComponent, {
+      width: '1600px',
+      data: {
+        userdata: data,
+        message: 'ویرایش کاربر ',
+        cancel: 'بستن',
+        submit: 'ذخیره اطلاعات',
+        submitFn: (data: User) => {
+          this.createUser(data);
+          dialog.close();
+        },
         cancelFn: () => {
           dialog.close();
         },
@@ -63,13 +105,17 @@ export class UsersTableComponent implements OnInit {
   }
 
   deleteUser(id: number) {
+    if (this.pending) return;
+    this.pending = true;
     this._userService.deleteUser(id).subscribe(
       () => {
+        this.pending = false;
         this.alertService.showSuccess('کاربر با موفقیت حذف شد');
         this.getUsersList();
       },
       () => {
-        this.alertService.showSuccess('خطایی هنگام حذف کاربر روی داد ');
+        this.pending = false;
+        this.alertService.showSuccess('خطایی هنگام حذف کاربر روی داده است ');
       }
     );
   }
